+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Not yet ready for use.
 
 ## v0.1.1 - TBD WIP
 - Minor credo refactors.
-- Brands: Dankort, GPN, LankaPay, Troy.
+- Brands: Dankort, GPN, LankaPay, Troy, UkrCard, Verve.
 
 ## v0.1.0 - 2023-08-05 WIP
 - Add guards to `card_brand`.